# docs/swagger.yaml
definitions:
...
  usercenter.APIInfo:
    properties:
      checkAuth:
        description: 是否检查权限
        type: boolean
      checkLogin:
        description: 是否需要登录
        type: boolean
      description:
        description: 描述
        type: string
      enable:
        description: 是否启用
        type: boolean
      group:
        description: 分组
        type: string
      id:
        description: UUID
        type: string
      method:
        description: |-
          方法 GET POST PUT 等等
          @inject_tag: validate:"required"
        type: string
      path:
        description: |-
          路径
          @inject_tag: validate:"required"
        type: string
    required:
    - method
    - path
    type: object
...
  usercenter.CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usercenter.DelRequest:
    properties:
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
...
  usercenter.GetMenuDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.MenuInfo'
      message:
        type: string
    type: object
...
  usercenter.MenuFunc:
    properties:
      hidden:
        type: boolean
      id:
        type: string
      menuFuncApis:
        items:
          $ref: '#/definitions/usercenter.MenuFuncApi'
        type: array
      menuID:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  usercenter.MenuFuncApi:
    properties:
      apiID:
        type: string
      apiInfo:
        $ref: '#/definitions/usercenter.APIInfo'
      id:
        type: string
      menuFuncID:
        type: string
    type: object
  usercenter.MenuInfo:
    properties:
      cache:
        type: boolean
      children:
        items:
          $ref: '#/definitions/usercenter.MenuInfo'
        type: array
      closeTab:
        type: boolean
      component:
        type: string
      defaultMenu:
        type: boolean
      hidden:
        type: boolean
      icon:
        type: string
      id:
        type: string
      level:
        type: integer
      menuFuncs:
        items:
          $ref: '#/definitions/usercenter.MenuFunc'
        type: array
      name:
        description: '@inject_tag: validate:"required"'
        type: string
      parameters:
        items:
          $ref: '#/definitions/usercenter.MenuParameter'
        type: array
      parentID:
        type: string
      path:
        type: string
      sort:
        type: integer
      title:
        type: string
    required:
    - name
    type: object
  usercenter.MenuParameter:
    properties:
      id:
        type: string
      key:
        type: string
      menuID:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
...
  usercenter.QueryMenuResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.MenuInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/core/auth/menu/add:
    post:
      consumes:
      - application/json
      description: 新增菜单
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.MenuInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增菜单
      tags:
      - 菜单管理
  /api/core/auth/menu/all:
    get:
      consumes:
      - application/json
      description: 查询所有菜单（Tree）
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryMenuResponse'
      summary: 查询所有菜单（Tree）
      tags:
      - 菜单管理
  /api/core/auth/menu/delete:
    delete:
      consumes:
      - application/json
      description: 软删除菜单
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除菜单
      tags:
      - 菜单管理
  /api/core/auth/menu/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetMenuDetailResponse'
      summary: 查询明细
      tags:
      - 菜单管理
  /api/core/auth/menu/export:
    get:
      consumes:
      - application/json
      description: 导出
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导出
      tags:
      - 菜单管理
  /api/core/auth/menu/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入
      parameters:
      - description: Bearer+空格+Token
        in: header
        name: authorization
        required: true
        type: string
      - description: 要上传的文件
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导入
      tags:
      - 菜单管理
  /api/core/auth/menu/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 路径
        in: query
        name: path
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: 显示名称
        in: query
        name: title
        type: string
      - description: 父ID
        in: query
        name: parentID
        type: string
      - description: 层级
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryMenuResponse'
      summary: 分页查询
      tags:
      - 菜单管理
  /api/core/auth/menu/update:
    put:
      consumes:
      - application/json
      description: 更新菜单
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.MenuInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新菜单
      tags:
      - 菜单管理
...
swagger: "2.0"